src/test/mlp/requirements.txt
flask
flask-socketio
python-socketio
opencv-python
numpy
torch
torchvision
pillow
colorama
termcolor
requests
pytest
python-socketio[client]
